Sprint 4 task 2

+++ db/migrate/20230307102857_add_points_to_employee_table.rb
@@ -1,5 +1,5 @@
 class AddPointsToEmployeeTable < ActiveRecord::Migration[6.1]
   def change
-    add_column :employees, :number_of_earned_points, :integer, default: 0
+    add_column :employees, :number_of_earned_points, :float, default: 0
   end
 end
